feat: pont 支持从 node_modules 读取自定义模板文件

+++ packages/pont-engine/src/utils/templateHelp.ts
@@ -6,6 +6,7 @@ import * as ts from 'typescript';
 import * as path from 'path';
 import { existsSync, writeFileSync } from 'fs-extra';
 import { LOCAL_DICT_DIR } from '../constants';
+import { Logger } from '../main/Logger';
 
 interface TemplateInfo {
   name: string;
@@ -14,31 +15,43 @@ interface TemplateInfo {
   defaultCode?: string;
 }
 
-export function getTemplate<T = any>(rootPath: string, templateInfo: TemplateInfo): T {
+export function getTemplate<T = any>(rootPath: string, templateInfo: TemplateInfo, originalPath?: string): T {
   const templateFileName = `${templateInfo.templatePath}.ts`;
+  const nodeModulesPath = `${rootPath}/node_modules/${originalPath}`;
 
-  if (!existsSync(templateFileName)) {
-    if (!templateInfo.defaultCode) return null;
-    writeFileSync(templateFileName, templateInfo.defaultCode);
+  try {
+    if (originalPath && existsSync(nodeModulesPath)) {
+      Logger.log(`[getTemplate ${templateInfo.templateType}] ${nodeModulesPath}`);
+      return require(nodeModulesPath);
+    }
+
+    if (!existsSync(templateFileName)) {
+      if (!templateInfo.defaultCode) return null;
+      writeFileSync(templateFileName, templateInfo.defaultCode);
+    }
+
+    const outDir = path.resolve(
+      rootPath,
+      `${LOCAL_DICT_DIR}/${templateInfo.templateType}/${templateInfo.name || 'default'}`
+    );
+    const outFile = `${outDir}${templateInfo.templatePath.split(rootPath)[1]}`;
+
+    const program = ts.createProgram([templateFileName], {
+      outDir,
+      rootDir: rootPath,
+      target: ts.ScriptTarget.ES2016,
+      module: ts.ModuleKind.CommonJS,
+      moduleResolution: ts.ModuleResolutionKind.NodeJs,
+      experimentalDecorators: true,
+      allowJs: true
+    });
+
+    program.emit();
+
+    Logger.log(`[getTemplate ${templateInfo.templateType}] ${outFile}`);
+    return require(outFile);
+  } catch (error) {
+    Logger.log(`[getTemplate error ${templateInfo.templateType}]`, error);
+    return null;
   }
-
-  const outDir = path.resolve(
-    rootPath,
-    `${LOCAL_DICT_DIR}/${templateInfo.templateType}/${templateInfo.name || 'default'}`
-  );
-  const outFile = `${outDir}${templateInfo.templatePath.split(rootPath)[1]}`;
-
-  const program = ts.createProgram([templateFileName], {
-    outDir,
-    rootDir: rootPath,
-    target: ts.ScriptTarget.ES2016,
-    module: ts.ModuleKind.CommonJS,
-    moduleResolution: ts.ModuleResolutionKind.NodeJs,
-    experimentalDecorators: true,
-    allowJs: true
-  });
-
-  program.emit();
-
-  return require(outFile);
 }
